hambedded.conf: add distribution config
Add master distribution configuration for hambedded. Currently, it is
based on poky distribution and it adds a few configuration on top of it
to help hamradio users easily use the distribution. IPK is used as a
package manager, and at least 4 virtual terminals are specified.

For testing purposes, yocto 1.4.1 package feeds are used.